Github workflow improvements - new steps to run locally with act/docker - cache - rename folders - fix proxystubs issue - 'apt update' before 'apt install' - nohup for the server - make upload-artifact optional
