--- RdkServicesTest/.gitignore
@@ -1,4 +1,0 @@
-/thunder
-/opt
-/valgrind_log
-/coverage
