+++ test/mocks/VideoDeviceMock.h
