+++ test/source/tr69hostif.py
@@ -1,5 +1,5 @@
-from http.server import BaseHTTPRequestHandler,HTTPServer
-import json
+from http.server import BaseHTTPRequestHandler, HTTPServer
+
 
 class S(BaseHTTPRequestHandler):
     def _set_headers(self):
@@ -11,9 +11,11 @@ class S(BaseHTTPRequestHandler):
         self._set_headers()
         self.wfile.write(b"{\"paramList\":[\"Device.X_CISCO_COM_LED.RedPwm\":123],\"success\":true}")
 
+
 def run(port=80):
     httpd = HTTPServer(('127.0.0.1', port), S)
     httpd.serve_forever()
+
 
 if __name__ == "__main__":
     run(10999)
