+++ test/tests/DeviceDiagnosticsTest.cpp
