+++ test/tests/SqliteStoreTest.cpp
